Crate ariel_os_sensors_gnss_time_ext

GNSS Time Extension

This extension to ariel-os-sensors allows you to read the time information from Samples produced by a compatible GNSS driver.

For users

Import the trait GnssTimeExt, you then access the time information on Samples generated by a compatible driver by using those functions:

GnssTimeExt::time_of_fix_timestamp(): UTC time in whole seconds since the UNIX epoch.

GnssTimeExt::time_of_fix_timestamp_nanos(): UTC time in nanoseconds since the UNIX epoch. Some sensors may have worse than nanosecond precision but this function will always return nanoseconds.

For implementors

Get parts with convert_datetime_to_parts(). You need to return a ReadingChannel with the label Label::OpaqueGnssTime containing the first part directly followed by a ReadingChannel with the label Label::Opaque containing the second part.

Enums

GnssTimeExtConvertError

Error returned when trying to convert the time to two parts.

GnssTimeExtError

Error returned when trying to access the time information on Samples coming from a GNSS sensor.

Traits

GnssTimeExt

Trait to access time information from Samples coming from a GNSS sensor.

Functions

convert_datetime_to_parts

Converts time to parts to be put into Samples.