ariel_os_sensors_gnss_time_ext

Function convert_datetime_to_parts

pub fn convert_datetime_to_parts(
    timestamp_nanos: i128,
) -> Result<(i32, i32), GnssTimeExtConvertError>

Converts time to parts to be put into Samples.

timestamp_nanos is an UTC timestamp as nanoseconds since the UNIX epoch.

The first element of the tuple is to be inserted in a ReadingChannel that has the label Label::OpaqueGnssTime. The second element must be inserted in a ReadingChannel right after the ReadingChannel of the first element and that has the label Label::Opaque.

You would return Samples like this:

// For demonstration purposes, In a normal implementation the Samples should be returned by
// `wait_for_reading()` implemented by the Sensor.
let sensor = SENSOR.init(MockSensor::new(ReadingChannels::from([
    ReadingChannel::new(
        Label::OpaqueGnssTime,
        0,
        ariel_os_sensors::MeasurementUnit::Second,
    ),
    ReadingChannel::new(Label::Opaque, -9, ariel_os_sensors::MeasurementUnit::Second),
])));

let timestamp_ns: i128 = 982107720621000000;
let (first, second) = convert_datetime_to_parts(timestamp_ns)?;
let samples = Samples::from_2(
    sensor,
    [
        Sample::new(
            first,
            ariel_os_sensors::sensor::SampleMetadata::UnknownAccuracy,
        ),
        Sample::new(
            second,
            ariel_os_sensors::sensor::SampleMetadata::UnknownAccuracy,
        ),
    ],
);

Errors

Returns an error when timestamp_nanos is either too far in the past or too far in the future.