ariel_os_sensors_gnss_time_ext

Trait GnssTimeExt

pub trait GnssTimeExt: Sealed {
    // Required methods
    fn time_of_fix_timestamp(&self) -> Result<i64, GnssTimeExtError>;
    fn time_of_fix_subsec_nanos(&self) -> Result<i64, GnssTimeExtError>;

    // Provided method
    fn time_of_fix_timestamp_nanos(&self) -> Result<i128, GnssTimeExtError> { ... }
}

Trait to access time information from Samples coming from a GNSS sensor.

Required Methods

fn time_of_fix_timestamp(&self) -> Result<i64, GnssTimeExtError>

Returns the UTC time in whole seconds since the UNIX epoch.

Errors

Returns errors if the reading is not compatible with this extension or one Sample returned an error.

fn time_of_fix_subsec_nanos(&self) -> Result<i64, GnssTimeExtError>

Returns the nanoseconds part of the time of the last fix.

Errors

Returns errors if the reading is not compatible with this extension or one Sample returned an error.

Provided Methods

fn time_of_fix_timestamp_nanos(&self) -> Result<i128, GnssTimeExtError>

Returns the time of the last fix in nanoseconds since the UNIX epoch.

Errors

Returns errors if the reading is not compatible with this extension or one Sample returned an error.